test(baseball): 테스트목적에따른 클래스 분류
